chore: Re-enable LIFO benchmarks in AsyncLockBenchmark
This commit removes the comments that were temporarily placed around the LIFO-based benchmark (`AcquireRelease_Concurrent_Lifo`) and the initialization of the `_lifoLock` field in the `AsyncLockBenchmark.cs` file.

This action re-enables the LIFO benchmarks to be included in the benchmark runs. This is likely done after a previous commit addressed a dependency issue (by adding a project reference in the benchmark runner's CSPROJ file) that might have necessitated their temporary exclusion.