debug: Trigger debug CI/CD workflow on 'Debug' branch
This commit modifies the `on.push.branches` setting in the `debug-ci-cd.yml` workflow. Instead of triggering on the `main` branch, this workflow will now be triggered specifically when changes are pushed to the `Debug` branch. This isolates the debug CI/CD pipeline, preventing it from running on regular development activities on the `main` branch and allowing for focused debugging efforts on the `Debug` branch.